packages/react-common-components/test/common/organisms/custom-arrows.test.tsx
import * as React from 'react';
import { mount } from 'enzyme';

import {
	CustomRightArrow,
	CustomLeftArrow
} from '../../../src/ui/common/organisms/oc-app-categories/custom-arrows';

describe('CategoryItem', () => {
	let leftArrow = mount(<CustomLeftArrow onClick={undefined} />);
	let rightArrow = mount(<CustomRightArrow onClick={undefined} />);

	it('should create', () => {
		expect(leftArrow).toBeTruthy();
		expect(rightArrow).toBeTruthy();
	});

	it('should have onClick prop', () => {
		expect(leftArrow.prop('onClick')).toEqual(undefined);
		leftArrow.setProps({ onClick: () => {} });
		leftArrow.update();
		expect(leftArrow.prop('onClick')).toBeTruthy();
		rightArrow.setProps({ onClick: () => {} });
		rightArrow.update();
		expect(rightArrow.prop('onClick')).toBeTruthy();
	});
});
